requirements.txt
pandas
seaborn
matplotlib
wordcloud
transformers
pytorch
ipywidgets
librosa
